refactor tidy pkg name
